serveur_rest/requirements.txt
flask
requests
pandas
cassandra-driver